safari-600.7-branch/LayoutTests/fast/dom/shadow/shadowroot-type-expected.txt

This test ensures ShadowRootType is correctly reflected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "AuthorShadowRoot" is internals.shadowRootType(shadowRootForDiv)
PASS "UserAgentShadowRoot" is internals.shadowRootType(shadowRootForInput)
PASS successfullyParsed is true

TEST COMPLETE
